- Implement UserAdminController with routes: /admin/users, /new, /edit, /save, /delete - Add JSPs: admin/users/list.jsp, admin/users/form.jsp - Update dashboard link to /admin/users - Keep optional redirect from /admin/users.jsp -> /admin/users - Ensure ADMIN-only access via AuthFilter

+++ src/main/java/com/pahanaedu/service/ItemService.java
@@ -19,7 +19,7 @@ public class ItemService {
     }
 
     public DeleteResult deleteSafe(int id) {
-        int uses = dao.countUsageInBills(id);
+        long uses = dao.countUsageInBills(id);   // <-- long
         if (uses > 0) return DeleteResult.IN_USE;
         if (uses < 0) return DeleteResult.ERROR;
         return dao.delete(id) ? DeleteResult.OK : DeleteResult.ERROR;
